Replace add_global_id with add_first_survey_user.
The name of the first survey user is taken from the user's first and last name.
However, because the registration form does not require the user to fill his
name, at the moment this name would be empty. This must be changed by adding
a name field to the registration form.

TODO: add 'name' field to the registration form