Add new tag (page_parents) to retrieve all parents of the current page.
This tag is used to get list of parent pages of the current page. The tag
is a template inclusion page, meaning that there is a separate template
that will retrieve the list. The template is also used to render the list
if required.

By default, it uses 'cms_extra/parents.html' template but it can be changed
by mentioning a different template in the tag argument like the following.

    {% page_parents "other_template_for_parent_pages.html" %}

The list of parent pages goes from the top most page to the direct parent
page. It can be reversed using the 'reversed' keyword in the Django's 'for'
template tag. Check the documentation here
http://docs.djangoproject.com/en/dev/ref/templates/builtins/#for