Backport localflavors for Belgium, Portugal and Sweden.
We are currently using Django version 1.1.1 which does not have localflavor
code for Belgium, Portugal and Sweden. Until we migrate to a more recent
version of Django, we backport the code for these countries to the Epiwork
code base (adapting it lightly where necessary) and patch fields.py to
refer to them.