Add functionality for administering an extra survey.
The survey should be given a title.  See gold-standard-contact.py

In settings.py, EXTRA_SURVEY should be set to the id of the desired survey
(not the title).  Setting it to False will disable this functionality.

An extra tab will appear with the title of the survey.  Clicking it will
bring up the survey.