Multiple countries in pollster_zip_codes and set in local_settings.py the user profile survey name
Compute the right zip code coords for the current user
when the same zip code appears for multiple countries in
pollster_zip_codes.

Alos moved in local_settings.py the shortname of the special user
profile survey toghether with the data name of the question used
to get its zip code and its country.